Move the context objects again for a Python2 import loop
Python2 considers that there's an import loop between
cli/__init__.py, cli/arguments/context_objects.py, and context.py.  Move
context_objects.py into utils/ to fix Python2
