+++ lib/ansible/utils/context_objects.py
