+++ test/units/executor/test_playbook_executor.py
@@ -22,10 +22,10 @@ __metaclass__ = type
 from units.compat import unittest
 from units.compat.mock import MagicMock
 
-from ansible.cli.arguments import context_objects as co
 from ansible.executor.playbook_executor import PlaybookExecutor
 from ansible.playbook import Playbook
 from ansible.template import Templar
+from ansible.utils import context_objects as co
 
 from units.mock.loader import DictDataLoader
