+++ test/units/playbook/test_play_context.py
@@ -13,12 +13,12 @@ import pytest
 
 from ansible import constants as C
 from ansible import context
-from ansible.cli.arguments import context_objects as co
 from ansible.cli.arguments import optparse_helpers as opt_help
-from units.compat import unittest
 from ansible.errors import AnsibleError, AnsibleParserError
 from ansible.module_utils.six.moves import shlex_quote
 from ansible.playbook.play_context import PlayContext
+from ansible.utils import context_objects as co
+from units.compat import unittest
 
 from units.mock.loader import DictDataLoader
